🚚 Move participant components into features/poll (RAL-1291)
- Move participants-provider, participant, participant-dropdown and
  new-participant-modal from components/ to features/poll/components/
- Remove new-participant-modal from the biome.json migration allowlist
- Temporarily allowlist poll-context.tsx, desktop-poll/participant-row-form.tsx,
  mobile-poll/poll-option.tsx and mobile-poll/poll-options.tsx: they now
  import the moved files and are themselves part of the final core cluster
