+++ apps/web/biome.json
@@ -125,12 +125,15 @@
         "src/components/forms/poll-settings.tsx",
         "src/components/layouts/poll-layout.tsx",
         "src/components/logo.tsx",
-        "src/components/new-participant-modal.tsx",
+        "src/components/poll-context.tsx",
         "src/components/poll/desktop-poll.tsx",
+        "src/components/poll/desktop-poll/participant-row-form.tsx",
         "src/components/poll/desktop-poll/participant-row.tsx",
         "src/components/poll/manage-poll.tsx",
         "src/components/poll/manage-poll/schedule-poll-dialog.tsx",
         "src/components/poll/mobile-poll.tsx",
+        "src/components/poll/mobile-poll/poll-option.tsx",
+        "src/components/poll/mobile-poll/poll-options.tsx",
         "src/components/poll/notification-toggle.tsx",
         "src/components/poll/poll-branding.tsx",
         "src/components/poll/poll-context-provider.tsx",
