+++ apps/web/src/app/[locale]/(optional-space)/poll/[urlId]/edit-options/page.tsx
@@ -8,12 +8,12 @@ import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
 import PollOptionsForm from "@/components/forms/poll-options-form";
 import { useModalContext } from "@/components/modal/modal-provider";
+import { useUpdatePollMutation } from "@/components/poll/mutations";
+import { usePoll } from "@/components/poll-context";
 import {
   filterParticipantsByVote,
   useParticipants,
-} from "@/components/participants-provider";
-import { useUpdatePollMutation } from "@/components/poll/mutations";
-import { usePoll } from "@/components/poll-context";
+} from "@/features/poll/components/participants-provider";
 import { Trans, useTranslation } from "@/i18n/client";
 import { dayjs } from "@/lib/dayjs";
 import { encodeDateOption } from "@/lib/utils/date-time-utils";
