+++ apps/web/src/components/poll/desktop-poll/participant-row.tsx
@@ -8,11 +8,14 @@ import { MoreHorizontalIcon } from "lucide-react";
 import type * as React from "react";
 
 import { OptimizedAvatarImage } from "@/components/optimized-avatar-image";
-import { Participant, ParticipantName } from "@/components/participant";
-import { ParticipantDropdown } from "@/components/participant-dropdown";
 import { usePoll } from "@/components/poll-context";
 import { useUser } from "@/components/user-provider";
 import { usePermissions } from "@/features/poll/client";
+import {
+  Participant,
+  ParticipantName,
+} from "@/features/poll/components/participant";
+import { ParticipantDropdown } from "@/features/poll/components/participant-dropdown";
 import { Trans, useTranslation } from "@/i18n/client";
 import type { Vote } from "@/trpc/client/types";
