+++ apps/web/src/components/poll/voting-form.tsx
@@ -9,15 +9,14 @@ import {
 import React from "react";
 import { FormProvider, useForm, useFormContext } from "react-hook-form";
 import * as z from "zod";
-
-import { NewParticipantForm } from "@/components/new-participant-modal";
-import { useParticipants } from "@/components/participants-provider";
 import {
   normalizeVotes,
   useEditToken,
   useUpdateParticipantMutation,
 } from "@/components/poll/mutations";
 import { usePermissions, usePoll, useRole } from "@/features/poll/client";
+import { NewParticipantForm } from "@/features/poll/components/new-participant-modal";
+import { useParticipants } from "@/features/poll/components/participants-provider";
 import { Trans } from "@/i18n/client";
 
 const formSchema = z.object({
