+++ apps/web/src/features/poll/components/participant-dropdown.tsx
