+++ apps/web/src/features/poll/components/participant.tsx
